worker/pyproject.toml
[project]
name = "vellic-worker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "arq>=0.25.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "cryptography>=42",
    "prometheus-client>=0.20.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "pytest-asyncio>=0.23.0", "pytest-cov>=5.0.0", "ruff>=0.4.0", "vellic-flags", "hypothesis>=6.0"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["."]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
vellic-flags = { path = "../packages/vellic_flags" }

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "vellic-flags",
]
